fix: prevent file watcher threads from slowing down JVM shutdown
FileWatcher was using an executor that creates non-daemon threads to watch
for file changes. This caused a delay in JVM shutdown, as it would
wait at least 60 seconds for all threads in the executor pool to be
evicted.

This change provides a custom thread factory that creates daemon threads
for the executor to prevent JVM shutdown delays.
